# openapi/scheduling-internal-v1.yaml
openapi: 3.0.3
info:
  title: AutoProduction Scheduling Internal API
  version: 1.0.0
  description: Internal API contract for scheduling console, approval workflow, publish/rollback and ops query.
servers:
  - url: /internal/v1
security:
  - bearerAuth: []
...
paths:
...
  /internal/replan-jobs:
    post:
      tags: [replan]
      summary: Trigger replan job
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplanJobRequest' }
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AcceptedResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Conflict:
      description: Business conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [request_id, code, message]
      properties:
        request_id: { type: string }
        code: { type: string }
        message: { type: string }
        retryable: { type: boolean }
...
    AcceptedResponse:
      type: object
      required: [request_id, accepted]
      properties:
        request_id: { type: string }
        accepted: { type: boolean }
        message: { type: string }
...
    ReplanJobRequest:
      type: object
      required: [request_id, trigger_type, scope_type, base_version_no, reason]
      properties:
        request_id: { type: string }
        trigger_type: { type: string, enum: [DELAY, PROGRESS_GAP, EQUIPMENT_DOWN, INSERT_ORDER] }
        scope_type: { type: string, enum: [LOCAL, GLOBAL] }
        base_version_no: { type: string }
        reason: { type: string }
...